feat: migrate testing framework into new testing module and improve test execution logging
