+++ arithmetization/src/test/java/net/consensys/linea/zktracer/module/add/AddTest.java
@@ -15,9 +15,9 @@
 
 package net.consensys.linea.zktracer.module.add;
 
+import net.consensys.linea.testing.BytecodeCompiler;
+import net.consensys.linea.testing.BytecodeRunner;
 import net.consensys.linea.zktracer.opcode.OpCode;
-import net.consensys.linea.zktracer.testing.BytecodeCompiler;
-import net.consensys.linea.zktracer.testing.BytecodeRunner;
 import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
