+++ arithmetization/src/test/java/net/consensys/linea/zktracer/module/exp/ExpTest.java
@@ -25,9 +25,9 @@ import java.util.List;
 import java.util.stream.Stream;
 
 import lombok.extern.slf4j.Slf4j;
+import net.consensys.linea.testing.BytecodeCompiler;
+import net.consensys.linea.testing.BytecodeRunner;
 import net.consensys.linea.zktracer.opcode.OpCode;
-import net.consensys.linea.zktracer.testing.BytecodeCompiler;
-import net.consensys.linea.zktracer.testing.BytecodeRunner;
 import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -234,6 +234,7 @@ public class ExpTest {
   }
 
   @Test
+  @Disabled
   void testExpUtils() {
     // ExpLog case
     log.debug("FFBlock");
