+++ arithmetization/src/test/java/net/consensys/linea/zktracer/module/mmu/MemoryTests.java
@@ -17,9 +17,9 @@ package net.consensys.linea.zktracer.module.mmu;
 
 import java.util.Random;
 
+import net.consensys.linea.testing.BytecodeCompiler;
+import net.consensys.linea.testing.BytecodeRunner;
 import net.consensys.linea.zktracer.opcode.OpCode;
-import net.consensys.linea.zktracer.testing.BytecodeCompiler;
-import net.consensys.linea.zktracer.testing.BytecodeRunner;
 import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
