+++ arithmetization/src/test/java/net/consensys/linea/zktracer/module/mod/ModTest.java
@@ -15,9 +15,9 @@
 
 package net.consensys.linea.zktracer.module.mod;
 
+import net.consensys.linea.testing.BytecodeCompiler;
+import net.consensys.linea.testing.BytecodeRunner;
 import net.consensys.linea.zktracer.opcode.OpCode;
-import net.consensys.linea.zktracer.testing.BytecodeCompiler;
-import net.consensys.linea.zktracer.testing.BytecodeRunner;
 import org.apache.tuweni.units.bigints.UInt256;
 import org.junit.jupiter.api.Test;
