+++ arithmetization/src/test/java/net/consensys/linea/zktracer/module/mxp/MxpTracerTest.java
@@ -22,9 +22,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
+import net.consensys.linea.testing.DynamicTests;
+import net.consensys.linea.testing.OpcodeCall;
 import net.consensys.linea.zktracer.module.Module;
-import net.consensys.linea.zktracer.testing.DynamicTests;
-import net.consensys.linea.zktracer.testing.OpcodeCall;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.units.bigints.UInt256;
 import org.junit.jupiter.api.DynamicTest;
