+++ testing/src/main/java/net/consensys/linea/testing/BytecodeRunner.java
@@ -13,7 +13,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.zktracer.testing;
+package net.consensys.linea.testing;
 
 import java.util.List;
 import java.util.Optional;
