+++ testing/src/main/java/net/consensys/linea/testing/DynamicTests.java
@@ -13,7 +13,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.zktracer.testing;
+package net.consensys.linea.testing;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
